pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "slideflow"
version = "0.0.0"
description = "Automated Google Slides presentation builder with charts and data replacements."
authors = [
    { name = "Joe Broadhead" },
    { name = "Tom Lovett" }
]
dependencies = [
    "typer[all]",
    "pydantic>=2.0",
    "pydantic[email]>=2.0",
    "pyyaml",
    "pandas",
    "numpy",
    "rich",
    "plotly",
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "httplib2",
    "gitpython",
    "dbt-core",
    "databricks-sql-connector",
]
requires-python = ">=3.12"

[project.optional-dependencies]
ai = ["openai", "google-generativeai"]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy",
]

[project.scripts]
slideflow = "slideflow.cli.main:app"

[tool.setuptools]
packages = ["slideflow"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true